requirements.txt
asgiref==3.9.2
Django==5.2.6
django-admin-rangefilter==0.13.3
django-smart-selects==1.7.2
sqlparse==0.5.3
tzdata==2025.2
